final commit

+++ src/Main.java
@@ -11,7 +11,7 @@ public class Main {
             while (it.hasNext()){
                 BestFirst.State i = it.next();
                 System.out.println(i);
-                if(!it.hasNext()) System.out.println(i.getG());
+                if(!it.hasNext()) System.out.println((int) i.getG());
             }
         }
         sc.close();
